+++ pywren_ibm_cloud/__init__.py
@@ -24,7 +24,7 @@ name = "pywren_ibm_cloud"
 def ibm_cf_executor(config=None, runtime=None, runtime_memory=None,
                     workers=None, region=None, storage_backend=None,
                     storage_backend_region=None, rabbitmq_monitor=None,
-                    log_level=None):
+                    remote_invoker=None, log_level=None):
     """
     Function executor for IBM Cloud Functions
     """
@@ -36,13 +36,14 @@ def ibm_cf_executor(config=None, runtime=None, runtime_memory=None,
         storage_backend=storage_backend,
         storage_backend_region=storage_backend_region,
         rabbitmq_monitor=rabbitmq_monitor,
+        remote_invoker=remote_invoker,
         log_level=log_level
     )
 
 
 def knative_executor(config=None, runtime=None, runtime_memory=None, workers=None,
                      region=None, storage_backend=None, storage_backend_region=None,
-                     rabbitmq_monitor=None, log_level=None):
+                     rabbitmq_monitor=None, remote_invoker=None, log_level=None):
     """
     Function executor for Knative
     """
@@ -54,6 +55,7 @@ def knative_executor(config=None, runtime=None, runtime_memory=None, workers=Non
         storage_backend=storage_backend,
         storage_backend_region=storage_backend_region,
         rabbitmq_monitor=rabbitmq_monitor,
+        remote_invoker=remote_invoker,
         log_level=log_level
     )
 
@@ -61,7 +63,7 @@ def knative_executor(config=None, runtime=None, runtime_memory=None, workers=Non
 def function_executor(config=None, runtime=None, runtime_memory=None,
                       workers=None, backend=None, region=None,
                       storage_backend=None, storage_backend_region=None,
-                      rabbitmq_monitor=None, log_level=None):
+                      rabbitmq_monitor=None, remote_invoker=None, log_level=None):
     """
     Generic function executor
     """
@@ -74,6 +76,7 @@ def function_executor(config=None, runtime=None, runtime_memory=None,
         storage_backend=storage_backend,
         storage_backend_region=storage_backend_region,
         rabbitmq_monitor=rabbitmq_monitor,
+        remote_invoker=remote_invoker,
         log_level=log_level
     )