+++ pywren_ibm_cloud/compute/backends/ibm_cf/ibm_cf.py
@@ -37,8 +37,8 @@ class IBMCloudFunctionsBackend:
         self.api_key = ibm_cf_config['regions'][self.region].get('api_key', None)
         self.iam_api_key = ibm_cf_config.get('iam_api_key', None)
 
-        logger.debug("Set IBM CF Namespace to {}".format(self.namespace))
-        logger.debug("Set IBM CF Endpoint to {}".format(self.endpoint))
+        logger.info("Set IBM CF Namespace to {}".format(self.namespace))
+        logger.info("Set IBM CF Endpoint to {}".format(self.endpoint))
 
         if self.api_key:
             self.cf_client = CloudFunctionsClient(endpoint=self.endpoint,
@@ -72,8 +72,7 @@ class IBMCloudFunctionsBackend:
             ibm_cf_config['token'] = token_manager._token
             ibm_cf_config['token_expiry_time'] = token_manager._expiry_time.strftime('%Y-%m-%d %H:%M:%S.%f%z')
 
-            self.cf_client = CloudFunctionsClient(region=self.region,
-                                                  endpoint=self.endpoint,
+            self.cf_client = CloudFunctionsClient(endpoint=self.endpoint,
                                                   namespace=self.namespace,
                                                   namespace_id=self.namespace_id,
                                                   token_manager=token_manager,
@@ -83,7 +82,7 @@ class IBMCloudFunctionsBackend:
                    'Region: {}'.format(__version__, self.namespace, self.region))
         if not self.log_level:
             print(log_msg)
-        logger.debug("IBM CF client created successfully")
+        logger.info("IBM CF client created successfully")
 
     def _format_action_name(self, runtime_name, runtime_memory):
         runtime_name = runtime_name.replace('/', '_').replace(':', '_')