+++ pywren_ibm_cloud/runtime/function_handler/__init__.py
@@ -1,1 +1,2 @@
 from .handler import function_handler
+from .invoker import function_invoker