+++ CHANGELOG.md
@@ -4,12 +4,23 @@
 
 ### Added
 - New invoker mechanism
+- New native remote invoker for ibm_cf
+- pywren-runtime clean command to delete all tmp data
+- capacity to limit the number of concurrent workers
+- architecture documentation
 
 ### Changed
-- 
+- Changed Internal data cleaner logic to delete only desired job
+- Updated ibm_cf Dockerfiles
+- Moved chunk min size from 1MB to 0MB
+- changed executor id format
+- Timeout waiting for functions to complete set to None by default
+- Updated ibm_cf base image requirements
 
 ### Fixed
-- 
+- Internal fixes
+- Fixed tests
+- Fixed pywren inside pywren function executions
 
 ## [v1.2.0]